Fix: for prompt downloading, let chrome decide the path
Issue #80

In the API doc, "item.filename" is stated as "Absolute local path". For prompted
download dialog, The last download location is not included in this path. So
pass no arguments to "suggest" to have Chrome automatically decide the path.

NOTE: for videos downloaded using this extension, this problem will persist as
there doesn't seem to be any API to get the last download location, so it will
always saved to "user's default Downloads directory".